requirements.txt
python-binance==1.0.29
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
python-dotenv==1.0.0
aiohttp==3.9.1
certifi>=2024.2.2 